+++ pkg/cmd/cli/client/client.go
@@ -30,7 +30,6 @@ func NewCommand() *cobra.Command {
 
 	c.AddCommand(
 		config.NewCommand(),
-		NewSetContextAsVeleroNamespaceCommand(),
 	)
 
 	return c
